# [ty] Speedup the `known_class_doesnt_fallback_to_unknown_unexpectedly_on_low_python_version` test