## Summary

This test takes >5s locally for me on `main`, which makes it our slowest unit test by far:

```
~/dev/ruff (main)⚡ % cargo test -p ty_python_semantic --lib -- known_class_doesnt_fallback_to_unknown_unexpectedly_on_low_python_version
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.09s
     Running unittests src/lib.rs (target/debug/deps/ty_python_semantic-346dd082aa08053a)

running 1 test
test types::class::tests::known_class_doesnt_fallback_to_unknown_unexpectedly_on_low_python_version ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 242 filtered out; finished in 5.30s
```

The slowness appears to be due to the fact that we're repeatedly changing the target Python version in a hot loop. This PR changes the test so that we only change the target Python version when it's actually necessary to do so, which brings the execution time of the test down to 0.39s for me locally:

```
~/dev/ruff (alex/speedup-test)⚡ % cargo test -p ty_python_semantic --lib -- known_class_doesnt_fallback_to_unknown_unexpectedly_on_low_python_version
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.12s
     Running unittests src/lib.rs (target/debug/deps/ty_python_semantic-346dd082aa08053a)

running 1 test
test types::class::tests::known_class_doesnt_fallback_to_unknown_unexpectedly_on_low_python_version ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 242 filtered out; finished in 0.39s
```

In turn, this brings the total execution time of `cargo test -p ty_python_semantic --lib` down from 6.67s to 0.54s.

## Test Plan

^See above